test(cmd): add test for version command
This PR sets up the tests which we'll use project
wide. We've changed the command to use a
constructor which makes it easier to test.

We've added the test file too with all acceptable
flag combos.

We could expand on this later but it's just the
version command so it's no big deal.